requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
psycopg2-binary==2.9.9
pymongo==4.6.0
python-dotenv==1.0.0
bcrypt==4.1.2
Werkzeug==3.0.1

